// src/app/model/data.model.ts
export class DataModel {
    title: string;
    description: string;
    issueTime: string;
    tokenFullName: string;
    symbol: string;
    totalSupply: string;
    circumSupply: string;
    website: string;
    explorer: string;
    deposite: string;
    withdrawals: string;
    trading: string;
    comment: string;
    media: string;
}